fixed object initializing: redudant deleted dynamic initializing block in ContentInitializer, some changes in fields; fixed types of fields (double to int)

+++ src/by/bntu/fitr/poisit/vauczeckaja/lab1/model/entity/content/Film.java
@@ -9,7 +9,7 @@ public class Film extends Content {
     public Film() {
     }
 
-    public Film(String name, String director, double budget, double gross) {
+    public Film(String name, String director, int budget, int gross) {
         this.name = name;
         this.creator = director;
         this.budget = budget;
